api - fetchinventory and clean up
